app/requirements.txt
Fflask==3.0.0
pytest==7.4.0
Werkzeug==3.0.1
flake8==7.0.0
